Use rubyarchhdrdir from RbConfig when available
If RbConfig has an entry for `rubyarchhdrdir` use it rather than joining
the platform-independent `rubyhdrdir` with the `arch`.  This is needed
for the vendor ruby on current Debian and Ubuntu systems where the
architecture dependent include files are in
`/usr/include/x86_64-linux-gnu/ruby-2.1.0`.  The old method would result
in the last two path components being switched.

Fall back to using the previous method if RbConfig doesn't define
`rubyarchhdrdir`.